refactored some of the preview related issues
