+++ Bet.AspNetCore.sln
@@ -69,6 +69,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bet.Extensions.ML.Sentiment
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bet.Extensions.AzureStorage", "src\Bet.Extensions.AzureStorage\Bet.Extensions.AzureStorage.csproj", "{E2DB931C-8B88-4870-A7D7-D51C411C141A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bet.AspNetCore.LetsEncrypt", "src\Bet.AspNetCore.LetsEncrypt\Bet.AspNetCore.LetsEncrypt.csproj", "{2D33AF22-F2D7-4862-861B-8C67537E0A18}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -151,6 +153,10 @@ Global
 		{E2DB931C-8B88-4870-A7D7-D51C411C141A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E2DB931C-8B88-4870-A7D7-D51C411C141A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E2DB931C-8B88-4870-A7D7-D51C411C141A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2D33AF22-F2D7-4862-861B-8C67537E0A18}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2D33AF22-F2D7-4862-861B-8C67537E0A18}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2D33AF22-F2D7-4862-861B-8C67537E0A18}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2D33AF22-F2D7-4862-861B-8C67537E0A18}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -175,6 +181,7 @@ Global
 		{03F2734D-F0FB-459F-963C-3BB249BB31D9} = {8382D62B-2BB8-4406-95FD-63AF167F5D65}
 		{40D0AD45-7772-4FA7-B566-E1F69FC180A0} = {8382D62B-2BB8-4406-95FD-63AF167F5D65}
 		{E2DB931C-8B88-4870-A7D7-D51C411C141A} = {8382D62B-2BB8-4406-95FD-63AF167F5D65}
+		{2D33AF22-F2D7-4862-861B-8C67537E0A18} = {8382D62B-2BB8-4406-95FD-63AF167F5D65}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A53F67D4-7CE9-4F03-8DD4-2C00E7AE2F46}
