+++ src/Bet.AspNetCore.Sample/Pages/BlobTest.cshtml
@@ -1,0 +1,8 @@
+﻿@page
+@model BlobTestModel
+@{
+    ViewData["Title"] = "BlobTestModel page";
+}
+
+<h1>BlobTest</h1>
+
